+++ src/components/charts/geo/ProportionalSymbolMap.tsx
@@ -223,7 +223,9 @@ export const ProportionalSymbolMap = forwardRef(function ProportionalSymbolMap<T
     size: [resolved.width, resolved.height],
     margin,
     enableHover: true,
-    tooltipContent: normalizeTooltip(tooltip) || defaultTooltip,
+    tooltipContent: tooltip === false
+      ? undefined
+      : (normalizeTooltip(tooltip) || defaultTooltip),
     ...(legend && { legend }),
     ...(legendInteraction && legendInteraction !== "none" && {
       legendHoverBehavior: legendState.onLegendHover,