+++ src/components/charts/ordinal/GroupedBarChart.tsx
@@ -176,7 +176,9 @@ export const GroupedBarChart = forwardRef(function GroupedBarChart<TDatum extend
     }),
     ...(title && { title }),
     ...(className && { className }),
-    tooltipContent: normalizeTooltip(tooltip) || defaultTooltipContent,
+    tooltipContent: tooltip === false
+      ? undefined
+      : (normalizeTooltip(tooltip) || defaultTooltipContent),
     ...((linkedHover || onObservation) && { customHoverBehavior }),
     ...(annotations && annotations.length > 0 && { annotations }),
     ...frameProps