+++ src/components/charts/ordinal/Histogram.tsx
@@ -156,7 +156,9 @@ export const Histogram = forwardRef(function Histogram<TDatum extends Record<str
     ...(legend && { legend }),
     ...(title && { title }),
     ...(className && { className }),
-    tooltipContent: normalizeTooltip(tooltip) || defaultTooltipContent,
+    tooltipContent: tooltip === false
+      ? undefined
+      : (normalizeTooltip(tooltip) || defaultTooltipContent),
     ...((linkedHover || onObservation) && { customHoverBehavior }),
     ...(annotations && annotations.length > 0 && { annotations }),
     ...frameProps